+++ release-notes.md
@@ -9,6 +9,16 @@
     - Cleaned up console output by hiding internal logging.
     - Made `run_all_tests.py` fully Python 3.x compatible.
 
+- **Refactored virtualenv management**: _(→ [Issue #2](https://github.com/martin-chatterjee/vfxtest/issues/2))_
+    - Added introspectable version.
+    - Now `pip install`'s specific `vfxtest` version into every virtualenv, to get set of matching dependencies.
+    - Now copies current `vfxtest.py` file to separate `PYTHONPATH` folder in dcc_settings.
+    - Now executes all subprocesses with the `-m vfxtest` args, instead of the absolute path to `vfxtest.py`.
+    - Introduced `use-environment` context setting for DCC contexts, specifying which Python environment they should use.
+      > **Warning**\
+      > This is a **breaking change** for DCC contexts. `use-environment` **must** be specified for all DCC contexts.
+    - Now uses `console_scripts` mechanism in `setup.py`.
+
 <br>
 
 ## `0.2.1` (09-Sep-2022)